+++ backend/tests/unit/security/test_security.py
@@ -11,13 +11,14 @@ import jwt
 
 from fastapi import HTTPException, Response
 from fastapi.security import SecurityScopes
+from src.config import Settings
 from src.security.security import (
     authenticate_password, get_authorization, get_current_user, create_access_token, generate_client_secret,
-    get_password_hash
+    get_password_hash, get_project_authorization
 )
 
 
-def test_create_access_token(settings):
+def test_create_access_token(settings: Settings):
     """ This tests a successfully created access token """
 
     payload_to_encode = {'sub': 'johndoe', 'scopes': ['read']}
@@ -133,6 +134,41 @@ def test_generate_client_secret():
     client_secret = generate_client_secret()
 
     assert len(client_secret) == 32
+
+
+def test_project_authorization_successful(
+    user_collection, analysis_collection, cpam0002_analysis_json, client_id_from_get_current_user
+):
+    """Shows a completely successful authentication process and what is required"""
+
+    analysis_collection.project_id_by_name = Mock(return_value=str(cpam0002_analysis_json["project_id"]))
+
+    database = {'user': user_collection, 'analysis': analysis_collection}
+
+    analysis_name = "CPAM0002"
+    authorization = get_project_authorization(
+        analysis_name, repositories=database, client_id=client_id_from_get_current_user
+    )
+
+    assert authorization is True
+
+
+@pytest.fixture(name="client_id_from_get_current_user")
+def fixture_client_id_from_get_current_user(response_with_fake_client_id_pyaload, settings):
+    """Fixture for retrieving the client_id from the response fixture according to the settings fixture"""
+    client_id = get_current_user(response_with_fake_client_id_pyaload, settings=settings)
+    return client_id
+
+
+@pytest.fixture(name="response_with_fake_client_id_pyaload")
+def fixture_reponse_with_fake_client_payload():
+    """Fixture for emulating a Reqeuest Response that includes a payload for the javascript web token to decode"""
+    payload = {
+        "sub": "fake-client-id",
+    }
+    jwt.decode = Mock(return_value=payload)
+
+    return Response("fake response", media_type="text/plain")
 
 
 @pytest.fixture(name="user_john_doe")